mqtt-subscriber/requirements.txt
paho-mqtt
redis
python-dateutil
